General optimisation work * preparatory work for optimisations * geometry 3D/2D faster volume/area computation for elements * reuse VMs more * fix problem which caused an openmp deadlock with too few dofs
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@2941 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ mesher/parallel_delaunay_3d.cpp
@@ -64,7 +64,7 @@ int ParallelDelaunayTree3D::getDomain(const DelaunayTetrahedron * tet) const
     
     for(size_t domain_index = 0 ; domain_index < domains.size() ;  domain_index++)
     {
-        if(domains[domain_index]->in(tet->getCenter()) && mesh == domain_index)
+        if(domains[domain_index]->in(tet->getCenter()) && mesh == (int)domain_index)
             return domain_index ;
     }
     return -1 ;
